requirements.txt
asgiref==3.2.3
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
Django==3.0.2
django-cors-headers==3.2.1
django-filter==2.2.0
djangorestframework==3.11.0
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.3
google-oauth2-tool==0.0.3
httplib2==0.17.0
idna==2.9
oauth2client==4.1.3
Pillow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pyOpenSSL==19.1.0
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.14.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.8
